separete the update modal into its own modal and them imported to home.js
