+++ src/components/Home.js
@@ -1,7 +1,8 @@
-import React, { useEffect, useState } from "react";
-import { Link } from "react-router-dom";
-import { useDispatch, useSelector } from "react-redux";
-import { deletePost, fetchPosts, updatePost } from "../redux/postSlice";
+import React, { useState, useEffect } from 'react';
+import { Link} from 'react-router-dom';
+import { useDispatch, useSelector } from 'react-redux';
+import { deletePost, fetchPosts, updatePost } from '../redux/postSlice';
+import EditModal from './EditeModal';
 
 const Home = () => {
   const dispatch = useDispatch();
@@ -9,11 +10,6 @@ const Home = () => {
   const loading = useSelector((state) => state.posts.loading);
   const [showEditModal, setShowEditModal] = useState(false);
   const [selectedPost, setSelectedPost] = useState(null);
-  const [editedFormData, setEditedFormData] = useState({
-    title: "",
-    author: "",
-    content: "",
-  });
 
   useEffect(() => {
     dispatch(fetchPosts());
@@ -25,32 +21,19 @@ const Home = () => {
 
   const handleUpdatePost = (post) => {
     setSelectedPost(post);
-    setEditedFormData({
-      title: post.title,
-      author: post.author,
-      content: post.content,
-    });
-
     setShowEditModal(true);
   };
 
   const handleEditModalClose = () => {
     setShowEditModal(false);
+    setSelectedPost(null); 
   };
 
-  const handleEditFormChange = (e) => {
-    const { name, value } = e.target;
-    setEditedFormData((prevData) => ({
-      ...prevData,
-      [name]: value,
-    }));
-  };
-
-  const handleEditFormSubmit = () => {
+  const handleEditFormSubmit = (editedFormData) => {
     dispatch(updatePost({ postId: selectedPost.id, formData: editedFormData }));
     setShowEditModal(false);
+    setSelectedPost(null); 
   };
-
 
   if (loading) {
     return <div>Loading...</div>;
@@ -72,55 +55,23 @@ const Home = () => {
             Author:
             <span>{post.author}</span>
           </p>
-          <p>
-            {post.content
-              ? `${post.content.slice(0, 50)}...`
-              : "No content available"}
-          </p>
-          <button onClick={() => handleDeletePost(post.id)}>Delete</button>
+          <p>{post.content ? `${post.content.slice(0, 50)}...` : 'No content available'}</p>
           <button onClick={() => handleUpdatePost(post)}>Update</button>
+          <button onClick={() => handleDeletePost(post.id)}>Delete</button>
         </div>
       ))}
 
       {showEditModal && (
-        <div className="modal">
-          <div className="modal-content">
-            <h2>Edit Post</h2>
-            <form onSubmit={(e) => e.preventDefault()}>
-              <label htmlFor="edit-title">Title:</label>
-              <input
-                id="edit-title"
-                type="text"
-                name="title"
-                value={editedFormData.title}
-                onChange={handleEditFormChange}
-              />
-
-              <label htmlFor="edit-author">Author:</label>
-              <input
-                id="edit-author"
-                type="text"
-                name="author"
-                value={editedFormData.author}
-                onChange={handleEditFormChange}
-              />
-
-              <label htmlFor="edit-content">Content:</label>
-              <textarea
-                id="edit-content"
-                name="content"
-                value={editedFormData.content}
-                onChange={handleEditFormChange}
-              />
-
-              <button onClick={handleEditFormSubmit}>Save Changes</button>
-              <button onClick={handleEditModalClose}>Cancel</button>
-            </form>
-          </div>
-        </div>
+        <EditModal
+          show={showEditModal}
+          onClose={handleEditModalClose}
+          onSubmit={handleEditFormSubmit}
+          post={selectedPost}
+        />
       )}
     </div>
   );
 };
 
 export default Home;
+
